requirements.txt
Flask==2.2.5
requests==2.31.0
authlib==0.15.6
python-dotenv==0.21.1
Flask-Session==0.4.0
typing-extensions==4.7.1
Werkzeug==2.2.3
